+++ xs/sources/mir/model_blocks.c
@@ -312,6 +312,20 @@ XsMirStatus xs_mir_block_rem_i32(XsMirBlock *block, XsMirValueId left, XsMirValu
                         error);
 }
 
+XsMirStatus xs_mir_block_and_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
+                                 XsMirError *error)
+{
+  return add_i32_binary(block, left, right, XS_MIR_INSTRUCTION_AND_I32, (XsMirType){.kind = XS_LIL_TYPE_I32}, result,
+                        error);
+}
+
+XsMirStatus xs_mir_block_or_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
+                                XsMirError *error)
+{
+  return add_i32_binary(block, left, right, XS_MIR_INSTRUCTION_OR_I32, (XsMirType){.kind = XS_LIL_TYPE_I32}, result,
+                        error);
+}
+
 XsMirStatus xs_mir_block_eq_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
                                 XsMirError *error)
 {