+++ xs/sources/mir/xlil_lowering.c
@@ -122,6 +122,22 @@ static XsMirStatus lower_instruction(const XsMirInstruction *instruction, XsLilB
                              &values[instruction->result], &lil_error);
     return map_lil_status(status, &lil_error, error);
   }
+  case XS_MIR_INSTRUCTION_AND_I32:
+  {
+    XsLilError lil_error = {0};
+    XsLilStatus status =
+        xs_lil_block_and_i32(block, values[instruction->operand_left], values[instruction->operand_right],
+                             &values[instruction->result], &lil_error);
+    return map_lil_status(status, &lil_error, error);
+  }
+  case XS_MIR_INSTRUCTION_OR_I32:
+  {
+    XsLilError lil_error = {0};
+    XsLilStatus status =
+        xs_lil_block_or_i32(block, values[instruction->operand_left], values[instruction->operand_right],
+                            &values[instruction->result], &lil_error);
+    return map_lil_status(status, &lil_error, error);
+  }
   case XS_MIR_INSTRUCTION_EQ_I32:
   {
     XsLilError lil_error = {0};